+++ test/express.static.js
@@ -1,7 +1,6 @@
 'use strict'
 
 var assert = require('assert')
-var Buffer = require('node:buffer').Buffer
 var express = require('..')
 var path = require('path')
 var request = require('supertest')