+++ test/express.urlencoded.js
@@ -2,7 +2,7 @@
 
 var assert = require('assert')
 var AsyncLocalStorage = require('async_hooks').AsyncLocalStorage
-var Buffer = require('node:buffer').Buffer
+
 var express = require('..')
 var request = require('supertest')