+++ History.md
@@ -3,6 +3,8 @@ unreleased
 
 * Remove `Object.setPrototypeOf` polyfill
 * cleanup: remove AsyncLocalStorage check from tests
+* cleanup: remove unnecessary require for global Buffer
+
 
 5.0.1 / 2024-10-08
 ==========